chore: move correlation id decorator to own package
