+++ src/SaanSoft.Cqrs.Decorator.EnsureCorrelationId/EnsureQueryHasCorrelationIdDecorator.cs
