+++ src/SaanSoft.Cqrs.Decorator.EnsureCorrelationId/ICorrelationIdProvider.cs
