+++ src/SaanSoft.Cqrs.Decorator.EnsureCorrelationId/SaanSoft.Cqrs.Decorator.EnsureCorrelationId.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>net7.0;net8.0</TargetFrameworks>
+    <IsPackable>true</IsPackable>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\SaanSoft.Cqrs.Decorator\SaanSoft.Cqrs.Decorator.csproj" />
+    <ProjectReference Include="..\SaanSoft.Cqrs.GuidIds\SaanSoft.Cqrs.GuidIds.csproj" />
+  </ItemGroup>
+
+</Project>
