+++ test/SaanSoft.Tests.Cqrs.Decorator.EnsureCorrelationId/GlobalUsings.cs
@@ -1,0 +1,5 @@
+global using FluentAssertions;
+global using SaanSoft.Cqrs.Decorator.EnsureCorrelationId.GuidIds;
+global using SaanSoft.Cqrs.GuidIds.Decorator;
+global using SaanSoft.Tests.Cqrs.Common.AutoFixture;
+global using SaanSoft.Tests.Cqrs.Common.TestModels;
