+++ test/SaanSoft.Tests.Cqrs.Decorator.EnsureCorrelationId/SaanSoft.Tests.Cqrs.Decorator.EnsureCorrelationId.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <IsPackable>false</IsPackable>
+    <IsTestProject>true</IsTestProject>
+    <NoWarn>CS1591;CS8602;CS8625;CS8618</NoWarn>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\SaanSoft.Cqrs.Decorator.EnsureCorrelationId\SaanSoft.Cqrs.Decorator.EnsureCorrelationId.csproj" />
+    <ProjectReference Include="..\SaanSoft.Tests.Cqrs.Decorator\SaanSoft.Tests.Cqrs.Decorator.csproj" />
+  </ItemGroup>
+
+</Project>
